requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Vision Models
timm>=0.9.0

# XAI Libraries
pytorch-grad-cam>=1.4.0
captum>=0.6.0
lime>=0.2.0

# Hilbert Curves
hilbertcurve>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
webdataset>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking (optional)
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.22.0

# Datasets
datasets>=2.12.0
huggingface-hub>=0.14.0
